@@ -1,10 +1,10 @@
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { Injectable, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { RouterModule } from '@angular/router';
-import { Routes } from '@angular/router/src/config';
+import { RouterModule, Routes } from '@angular/router';
 import { HelgolandCachingModule } from '@helgoland/caching';
 import { DatasetApiInterface, Settings, SettingsService, SplittedDataDatasetApiInterface } from '@helgoland/core';
+import { VocabNercLabelMapperModule } from '@helgoland/depiction';
 import { JsonFavoriteExporterService } from '@helgoland/favorite';
 import {
   NgbAccordionModule,
@@ -72,6 +72,7 @@ export class ExtendedSettingsService extends SettingsService<Settings> {
     NgbDropdownModule,
     NgbDatepickerModule,
     NgbTimepickerModule,
+    VocabNercLabelMapperModule,
     RouterModule.forRoot(
       baseRoutes
     )
