@@ -64,6 +64,10 @@ export class TimeseriesDiagramComponent implements OnInit {
         this.timeseriesService.removeDataset(internalId);
     }
 
+    public deleteAllTs() {
+        this.timeseriesService.removeAllDatasets();
+    }
+
     public selectTimeseries(selected: boolean, internalId: string) {
         if (selected) {
             this.selectedIds.push(internalId);
@@ -113,7 +117,7 @@ export class TimeseriesDiagramComponent implements OnInit {
     }
 
     public showGeometry(geometry: GeoJSON.GeoJsonObject) {
-        const ref = this.modalService.open(ModalGeometryViewerComponent);
+        const ref = this.modalService.open(ModalGeometryViewerComponent, { size: 'lg' });
         (ref.componentInstance as ModalGeometryViewerComponent).geometry = geometry;
     }
 
